requirements.txt
cssselect==0.9.1
lxml==3.4.4
pyquery==1.2.9
requests==2.7.0
wheel==0.24.0
xively-python==0.1.0rc1
